+++ assets/css/screenplay-container.scss
@@ -25,7 +25,4 @@
       flex-grow: 1;
     }
   }
-  .page-content__hidden {
-    display: none;
-  }
 }
