+++ assets/js/components/App.tsx
@@ -15,8 +15,14 @@ class AppRoutes extends React.Component {
             path="/emergency-takeover"
             element={<OutfrontTakeoverTool />}
           ></Route>
-          <Route path="/dashboard" element={<Dashboard />}></Route>
-          <Route path="/dashboard/alerts" element={<Dashboard />}></Route>
+          <Route
+            path="/dashboard"
+            element={<Dashboard page="places" />}
+          ></Route>
+          <Route
+            path="/dashboard/alerts"
+            element={<Dashboard page="alerts" />}
+          ></Route>
         </Routes>
       </React.Suspense>
     );
